+++ tests/Router/TestMiddleware.php
@@ -7,5 +7,8 @@ class TestMiddleware extends AbstractMiddleware
   public function handle()
   {
     $_SERVER['middleware'] = 'oke';
+    self::$last++;
   }
+
+  public static $last = 0;
 }