+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+###  Added
+- Prevent duplicate middleware call.
+- Added method (`Route::middleware`) to call middleware individualy.
+
 ### Changed
 - Change `Router::middleware` param form middleware object to middlewar class-name.
 - `RouterGroup::group` retrun as callback param.